+++ warp/tests/test_builtins_resolution.py
@@ -107,17 +107,32 @@ class TestBuiltinsResolution(unittest.TestCase):
         expected = 0.94248880193169748409
 
         result = wp.sin(wp.float64(value))
-        self.assertAlmostEqual(result, expected, places=12)
+        self.assertIsInstance(result, wp.float64)
+        self.assertAlmostEqual(float(result), expected, places=12)
 
         result = wp.sin(wp.float32(value))
+        self.assertIsInstance(result, float)
         self.assertNotAlmostEqual(result, expected, places=12)
         self.assertAlmostEqual(result, expected, places=5)
 
         result = wp.sin(wp.float16(value))
-        self.assertNotAlmostEqual(result, expected, places=5)
-        self.assertAlmostEqual(result, expected, places=1)
+        self.assertIsInstance(result, wp.float16)
+        self.assertNotAlmostEqual(float(result), expected, places=5)
+        self.assertAlmostEqual(float(result), expected, places=1)
 
         self.assertEqual(wp.sin(value), wp.sin(wp.float32(value)))
+
+    def test_legacy_scalar_return_types(self):
+        old_setting = wp.config.legacy_scalar_return_types
+
+        try:
+            wp.config.legacy_scalar_return_types = True
+
+            self.assertIsInstance(wp.sin(wp.float16(1.23)), float)
+            self.assertIsInstance(wp.sin(wp.float64(1.23)), float)
+            self.assertIsInstance(wp.invert(wp.int16(-123)), int)
+        finally:
+            wp.config.legacy_scalar_return_types = old_setting
 
     def test_int_int_args_overflow(self):
         value = -1234567890
@@ -139,45 +154,54 @@ class TestBuiltinsResolution(unittest.TestCase):
         expected = 5.78999999999999914735
 
         result = wp.trace(wp.mat22d(*values))
-        self.assertAlmostEqual(result, expected, places=12)
+        self.assertIsInstance(result, wp.float64)
+        self.assertAlmostEqual(float(result), expected, places=12)
 
         result = wp.trace(wp.mat22f(*values))
+        self.assertIsInstance(result, float)
         self.assertNotAlmostEqual(result, expected, places=12)
         self.assertAlmostEqual(result, expected, places=5)
 
         result = wp.trace(wp.mat22h(*values))
-        self.assertNotAlmostEqual(result, expected, places=5)
-        self.assertAlmostEqual(result, expected, places=1)
+        self.assertIsInstance(result, wp.float16)
+        self.assertNotAlmostEqual(float(result), expected, places=5)
+        self.assertAlmostEqual(float(result), expected, places=1)
 
     def test_mat33_arg_precision(self):
         values = (1.23, 2.34, 3.45, 4.56, 5.67, 6.78, 7.89, 8.90, 9.01)
         expected = 15.91000000000000014211
 
         result = wp.trace(wp.mat33d(*values))
-        self.assertAlmostEqual(result, expected, places=12)
+        self.assertIsInstance(result, wp.float64)
+        self.assertAlmostEqual(float(result), expected, places=12)
 
         result = wp.trace(wp.mat33f(*values))
+        self.assertIsInstance(result, float)
         self.assertNotAlmostEqual(result, expected, places=12)
         self.assertAlmostEqual(result, expected, places=5)
 
         result = wp.trace(wp.mat33h(*values))
-        self.assertNotAlmostEqual(result, expected, places=5)
-        self.assertAlmostEqual(result, expected, places=1)
+        self.assertIsInstance(result, wp.float16)
+        self.assertNotAlmostEqual(float(result), expected, places=5)
+        self.assertAlmostEqual(float(result), expected, places=1)
 
     def test_mat44_arg_precision(self):
         values = (1.23, 2.34, 3.45, 4.56, 5.67, 6.78, 7.89, 8.90, 9.01, 10.12, 11.23, 12.34, 13.45, 14.56, 15.67, 16.78)
         expected = 36.02000000000000312639
 
         result = wp.trace(wp.mat44d(*values))
-        self.assertAlmostEqual(result, expected, places=12)
+        self.assertIsInstance(result, wp.float64)
+        self.assertAlmostEqual(float(result), expected, places=12)
 
         result = wp.trace(wp.mat44f(*values))
+        self.assertIsInstance(result, float)
         self.assertNotAlmostEqual(result, expected, places=12)
         self.assertAlmostEqual(result, expected, places=5)
 
         result = wp.trace(wp.mat44h(*values))
-        self.assertNotAlmostEqual(result, expected, places=5)
-        self.assertAlmostEqual(result, expected, places=1)
+        self.assertIsInstance(result, wp.float16)
+        self.assertNotAlmostEqual(float(result), expected, places=5)
+        self.assertAlmostEqual(float(result), expected, places=1)
 
     def test_mat22_mat22_args_precision(self):
         a_values = (0.12, 1.23, 0.12, 1.23)
@@ -185,15 +209,18 @@ class TestBuiltinsResolution(unittest.TestCase):
         expected = 0.59039999999999992486
 
         result = wp.ddot(wp.mat22d(*a_values), wp.mat22d(*b_values))
-        self.assertAlmostEqual(result, expected, places=12)
+        self.assertIsInstance(result, wp.float64)
+        self.assertAlmostEqual(float(result), expected, places=12)
 
         result = wp.ddot(wp.mat22f(*a_values), wp.mat22f(*b_values))
+        self.assertIsInstance(result, float)
         self.assertNotAlmostEqual(result, expected, places=12)
         self.assertAlmostEqual(result, expected, places=5)
 
         result = wp.ddot(wp.mat22h(*a_values), wp.mat22h(*b_values))
-        self.assertNotAlmostEqual(result, expected, places=5)
-        self.assertAlmostEqual(result, expected, places=1)
+        self.assertIsInstance(result, wp.float16)
+        self.assertNotAlmostEqual(float(result), expected, places=5)
+        self.assertAlmostEqual(float(result), expected, places=1)
 
     def test_mat33_mat33_args_precision(self):
         a_values = (0.12, 1.23, 2.34, 0.12, 1.23, 2.34, 0.12, 1.23, 2.34)
@@ -201,15 +228,18 @@ class TestBuiltinsResolution(unittest.TestCase):
         expected = 6.22350000000000047606
 
         result = wp.ddot(wp.mat33d(*a_values), wp.mat33d(*b_values))
-        self.assertAlmostEqual(result, expected, places=12)
+        self.assertIsInstance(result, wp.float64)
+        self.assertAlmostEqual(float(result), expected, places=12)
 
         result = wp.ddot(wp.mat33f(*a_values), wp.mat33f(*b_values))
+        self.assertIsInstance(result, float)
         self.assertNotAlmostEqual(result, expected, places=12)
         self.assertAlmostEqual(result, expected, places=5)
 
         result = wp.ddot(wp.mat33h(*a_values), wp.mat33h(*b_values))
-        self.assertNotAlmostEqual(result, expected, places=5)
-        self.assertAlmostEqual(result, expected, places=1)
+        self.assertIsInstance(result, wp.float16)
+        self.assertNotAlmostEqual(float(result), expected, places=5)
+        self.assertAlmostEqual(float(result), expected, places=1)
 
     def test_mat44_mat44_args(self):
         a_values = (0.12, 1.23, 2.34, 3.45, 0.12, 1.23, 2.34, 3.45, 0.12, 1.23, 2.34, 3.45, 0.12, 1.23, 2.34, 3.45)
@@ -217,15 +247,18 @@ class TestBuiltinsResolution(unittest.TestCase):
         expected = 26.33760000000000189857
 
         result = wp.ddot(wp.mat44d(*a_values), wp.mat44d(*b_values))
-        self.assertAlmostEqual(result, expected, places=12)
+        self.assertIsInstance(result, wp.float64)
+        self.assertAlmostEqual(float(result), expected, places=12)
 
         result = wp.ddot(wp.mat44f(*a_values), wp.mat44f(*b_values))
+        self.assertIsInstance(result, float)
         self.assertNotAlmostEqual(result, expected, places=12)
         self.assertAlmostEqual(result, expected, places=5)
 
         result = wp.ddot(wp.mat44h(*a_values), wp.mat44h(*b_values))
-        self.assertNotAlmostEqual(result, expected, places=5)
-        self.assertAlmostEqual(result, expected, places=1)
+        self.assertIsInstance(result, wp.float16)
+        self.assertNotAlmostEqual(float(result), expected, places=5)
+        self.assertAlmostEqual(float(result), expected, places=1)
 
     def test_mat22_float_args_precision(self):
         a_values = (1.23, 2.34, 3.45, 4.56)
@@ -236,10 +269,10 @@ class TestBuiltinsResolution(unittest.TestCase):
         expected_11 = 0.54719999999999990870
 
         result = wp.mul(wp.mat22d(*a_values), wp.float64(b_value))
-        self.assertAlmostEqual(result[0][0], expected_00, places=12)
-        self.assertAlmostEqual(result[0][1], expected_01, places=12)
-        self.assertAlmostEqual(result[1][0], expected_10, places=12)
-        self.assertAlmostEqual(result[1][1], expected_11, places=12)
+        self.assertAlmostEqual(float(result[0][0]), expected_00, places=12)
+        self.assertAlmostEqual(float(result[0][1]), expected_01, places=12)
+        self.assertAlmostEqual(float(result[1][0]), expected_10, places=12)
+        self.assertAlmostEqual(float(result[1][1]), expected_11, places=12)
 
         result = wp.mul(wp.mat22f(*a_values), wp.float32(b_value))
         self.assertNotAlmostEqual(result[0][0], expected_00, places=12)
@@ -252,14 +285,14 @@ class TestBuiltinsResolution(unittest.TestCase):
         self.assertAlmostEqual(result[1][1], expected_11, places=5)
 
         result = wp.mul(wp.mat22h(*a_values), wp.float16(b_value))
-        self.assertNotAlmostEqual(result[0][0], expected_00, places=5)
-        self.assertNotAlmostEqual(result[0][1], expected_01, places=5)
-        self.assertNotAlmostEqual(result[1][0], expected_10, places=5)
-        self.assertNotAlmostEqual(result[1][1], expected_11, places=5)
-        self.assertAlmostEqual(result[0][0], expected_00, places=1)
-        self.assertAlmostEqual(result[0][1], expected_01, places=1)
-        self.assertAlmostEqual(result[1][0], expected_10, places=1)
-        self.assertAlmostEqual(result[1][1], expected_11, places=1)
+        self.assertNotAlmostEqual(float(result[0][0]), expected_00, places=5)
+        self.assertNotAlmostEqual(float(result[0][1]), expected_01, places=5)
+        self.assertNotAlmostEqual(float(result[1][0]), expected_10, places=5)
+        self.assertNotAlmostEqual(float(result[1][1]), expected_11, places=5)
+        self.assertAlmostEqual(float(result[0][0]), expected_00, places=1)
+        self.assertAlmostEqual(float(result[0][1]), expected_01, places=1)
+        self.assertAlmostEqual(float(result[1][0]), expected_10, places=1)
+        self.assertAlmostEqual(float(result[1][1]), expected_11, places=1)
 
     def test_mat33_float_args_precision(self):
         a_values = (1.23, 2.34, 3.45, 4.56, 5.67, 6.78, 7.89, 8.90, 9.01)
@@ -275,15 +308,15 @@ class TestBuiltinsResolution(unittest.TestCase):
         expected_22 = 1.08119999999999993889
 
         result = wp.mul(wp.mat33d(*a_values), wp.float64(b_value))
-        self.assertAlmostEqual(result[0][0], expected_00, places=12)
-        self.assertAlmostEqual(result[0][1], expected_01, places=12)
-        self.assertAlmostEqual(result[0][2], expected_02, places=12)
-        self.assertAlmostEqual(result[1][0], expected_10, places=12)
-        self.assertAlmostEqual(result[1][1], expected_11, places=12)
-        self.assertAlmostEqual(result[1][2], expected_12, places=12)
-        self.assertAlmostEqual(result[2][0], expected_20, places=12)
-        self.assertAlmostEqual(result[2][1], expected_21, places=12)
-        self.assertAlmostEqual(result[2][2], expected_22, places=12)
+        self.assertAlmostEqual(float(result[0][0]), expected_00, places=12)
+        self.assertAlmostEqual(float(result[0][1]), expected_01, places=12)
+        self.assertAlmostEqual(float(result[0][2]), expected_02, places=12)
+        self.assertAlmostEqual(float(result[1][0]), expected_10, places=12)
+        self.assertAlmostEqual(float(result[1][1]), expected_11, places=12)
+        self.assertAlmostEqual(float(result[1][2]), expected_12, places=12)
+        self.assertAlmostEqual(float(result[2][0]), expected_20, places=12)
+        self.assertAlmostEqual(float(result[2][1]), expected_21, places=12)
+        self.assertAlmostEqual(float(result[2][2]), expected_22, places=12)
 
         result = wp.mul(wp.mat33f(*a_values), wp.float32(b_value))
         self.assertNotAlmostEqual(result[0][0], expected_00, places=12)
@@ -306,24 +339,24 @@ class TestBuiltinsResolution(unittest.TestCase):
         self.assertAlmostEqual(result[2][2], expected_22, places=5)
 
         result = wp.mul(wp.mat33h(*a_values), wp.float16(b_value))
-        self.assertNotAlmostEqual(result[0][0], expected_00, places=5)
-        self.assertNotAlmostEqual(result[0][1], expected_01, places=5)
-        self.assertNotAlmostEqual(result[0][2], expected_02, places=5)
-        self.assertNotAlmostEqual(result[1][0], expected_10, places=5)
-        self.assertNotAlmostEqual(result[1][1], expected_11, places=5)
-        self.assertNotAlmostEqual(result[1][2], expected_12, places=5)
-        self.assertNotAlmostEqual(result[2][0], expected_20, places=5)
-        self.assertNotAlmostEqual(result[2][1], expected_21, places=5)
-        self.assertNotAlmostEqual(result[2][2], expected_22, places=5)
-        self.assertAlmostEqual(result[0][0], expected_00, places=1)
-        self.assertAlmostEqual(result[0][1], expected_01, places=1)
-        self.assertAlmostEqual(result[0][2], expected_02, places=1)
-        self.assertAlmostEqual(result[1][0], expected_10, places=1)
-        self.assertAlmostEqual(result[1][1], expected_11, places=1)
-        self.assertAlmostEqual(result[1][2], expected_12, places=1)
-        self.assertAlmostEqual(result[2][0], expected_20, places=1)
-        self.assertAlmostEqual(result[2][1], expected_21, places=1)
-        self.assertAlmostEqual(result[2][2], expected_22, places=1)
+        self.assertNotAlmostEqual(float(result[0][0]), expected_00, places=5)
+        self.assertNotAlmostEqual(float(result[0][1]), expected_01, places=5)
+        self.assertNotAlmostEqual(float(result[0][2]), expected_02, places=5)
+        self.assertNotAlmostEqual(float(result[1][0]), expected_10, places=5)
+        self.assertNotAlmostEqual(float(result[1][1]), expected_11, places=5)
+        self.assertNotAlmostEqual(float(result[1][2]), expected_12, places=5)
+        self.assertNotAlmostEqual(float(result[2][0]), expected_20, places=5)
+        self.assertNotAlmostEqual(float(result[2][1]), expected_21, places=5)
+        self.assertNotAlmostEqual(float(result[2][2]), expected_22, places=5)
+        self.assertAlmostEqual(float(result[0][0]), expected_00, places=1)
+        self.assertAlmostEqual(float(result[0][1]), expected_01, places=1)
+        self.assertAlmostEqual(float(result[0][2]), expected_02, places=1)
+        self.assertAlmostEqual(float(result[1][0]), expected_10, places=1)
+        self.assertAlmostEqual(float(result[1][1]), expected_11, places=1)
+        self.assertAlmostEqual(float(result[1][2]), expected_12, places=1)
+        self.assertAlmostEqual(float(result[2][0]), expected_20, places=1)
+        self.assertAlmostEqual(float(result[2][1]), expected_21, places=1)
+        self.assertAlmostEqual(float(result[2][2]), expected_22, places=1)
 
     def test_mat44_float_args_precision(self):
         a_values = (
@@ -363,22 +396,22 @@ class TestBuiltinsResolution(unittest.TestCase):
         expected_33 = 2.01360000000000027853
 
         result = wp.mul(wp.mat44d(*a_values), wp.float64(b_value))
-        self.assertAlmostEqual(result[0][0], expected_00, places=12)
-        self.assertAlmostEqual(result[0][1], expected_01, places=12)
-        self.assertAlmostEqual(result[0][2], expected_02, places=12)
-        self.assertAlmostEqual(result[0][3], expected_03, places=12)
-        self.assertAlmostEqual(result[1][0], expected_10, places=12)
-        self.assertAlmostEqual(result[1][1], expected_11, places=12)
-        self.assertAlmostEqual(result[1][2], expected_12, places=12)
-        self.assertAlmostEqual(result[1][3], expected_13, places=12)
-        self.assertAlmostEqual(result[2][0], expected_20, places=12)
-        self.assertAlmostEqual(result[2][1], expected_21, places=12)
-        self.assertAlmostEqual(result[2][2], expected_22, places=12)
-        self.assertAlmostEqual(result[2][3], expected_23, places=12)
-        self.assertAlmostEqual(result[3][0], expected_30, places=12)
-        self.assertAlmostEqual(result[3][1], expected_31, places=12)
-        self.assertAlmostEqual(result[3][2], expected_32, places=12)
-        self.assertAlmostEqual(result[3][3], expected_33, places=12)
+        self.assertAlmostEqual(float(result[0][0]), expected_00, places=12)
+        self.assertAlmostEqual(float(result[0][1]), expected_01, places=12)
+        self.assertAlmostEqual(float(result[0][2]), expected_02, places=12)
+        self.assertAlmostEqual(float(result[0][3]), expected_03, places=12)
+        self.assertAlmostEqual(float(result[1][0]), expected_10, places=12)
+        self.assertAlmostEqual(float(result[1][1]), expected_11, places=12)
+        self.assertAlmostEqual(float(result[1][2]), expected_12, places=12)
+        self.assertAlmostEqual(float(result[1][3]), expected_13, places=12)
+        self.assertAlmostEqual(float(result[2][0]), expected_20, places=12)
+        self.assertAlmostEqual(float(result[2][1]), expected_21, places=12)
+        self.assertAlmostEqual(float(result[2][2]), expected_22, places=12)
+        self.assertAlmostEqual(float(result[2][3]), expected_23, places=12)
+        self.assertAlmostEqual(float(result[3][0]), expected_30, places=12)
+        self.assertAlmostEqual(float(result[3][1]), expected_31, places=12)
+        self.assertAlmostEqual(float(result[3][2]), expected_32, places=12)
+        self.assertAlmostEqual(float(result[3][3]), expected_33, places=12)
 
         result = wp.mul(wp.mat44f(*a_values), wp.float32(b_value))
         self.assertNotAlmostEqual(result[0][0], expected_00, places=12)
@@ -415,53 +448,56 @@ class TestBuiltinsResolution(unittest.TestCase):
         self.assertAlmostEqual(result[3][3], expected_33, places=5)
 
         result = wp.mul(wp.mat44h(*a_values), wp.float16(b_value))
-        self.assertNotAlmostEqual(result[0][0], expected_00, places=5)
-        self.assertNotAlmostEqual(result[0][1], expected_01, places=5)
-        self.assertNotAlmostEqual(result[0][2], expected_02, places=5)
-        self.assertNotAlmostEqual(result[0][3], expected_03, places=5)
-        self.assertNotAlmostEqual(result[1][0], expected_10, places=5)
-        self.assertNotAlmostEqual(result[1][1], expected_11, places=5)
-        self.assertNotAlmostEqual(result[1][2], expected_12, places=5)
-        self.assertNotAlmostEqual(result[1][3], expected_13, places=5)
-        self.assertNotAlmostEqual(result[2][0], expected_20, places=5)
-        self.assertNotAlmostEqual(result[2][1], expected_21, places=5)
-        self.assertNotAlmostEqual(result[2][2], expected_22, places=5)
-        self.assertNotAlmostEqual(result[2][3], expected_23, places=5)
-        self.assertNotAlmostEqual(result[3][0], expected_30, places=5)
-        self.assertNotAlmostEqual(result[3][1], expected_31, places=5)
-        self.assertNotAlmostEqual(result[3][2], expected_32, places=5)
-        self.assertNotAlmostEqual(result[3][3], expected_33, places=5)
-        self.assertAlmostEqual(result[0][0], expected_00, places=1)
-        self.assertAlmostEqual(result[0][1], expected_01, places=1)
-        self.assertAlmostEqual(result[0][2], expected_02, places=1)
-        self.assertAlmostEqual(result[0][3], expected_03, places=1)
-        self.assertAlmostEqual(result[1][0], expected_10, places=1)
-        self.assertAlmostEqual(result[1][1], expected_11, places=1)
-        self.assertAlmostEqual(result[1][2], expected_12, places=1)
-        self.assertAlmostEqual(result[1][3], expected_13, places=1)
-        self.assertAlmostEqual(result[2][0], expected_20, places=1)
-        self.assertAlmostEqual(result[2][1], expected_21, places=1)
-        self.assertAlmostEqual(result[2][2], expected_22, places=1)
-        self.assertAlmostEqual(result[2][3], expected_23, places=1)
-        self.assertAlmostEqual(result[3][0], expected_30, places=1)
-        self.assertAlmostEqual(result[3][1], expected_31, places=1)
-        self.assertAlmostEqual(result[3][2], expected_32, places=1)
-        self.assertAlmostEqual(result[3][3], expected_33, places=1)
+        self.assertNotAlmostEqual(float(result[0][0]), expected_00, places=5)
+        self.assertNotAlmostEqual(float(result[0][1]), expected_01, places=5)
+        self.assertNotAlmostEqual(float(result[0][2]), expected_02, places=5)
+        self.assertNotAlmostEqual(float(result[0][3]), expected_03, places=5)
+        self.assertNotAlmostEqual(float(result[1][0]), expected_10, places=5)
+        self.assertNotAlmostEqual(float(result[1][1]), expected_11, places=5)
+        self.assertNotAlmostEqual(float(result[1][2]), expected_12, places=5)
+        self.assertNotAlmostEqual(float(result[1][3]), expected_13, places=5)
+        self.assertNotAlmostEqual(float(result[2][0]), expected_20, places=5)
+        self.assertNotAlmostEqual(float(result[2][1]), expected_21, places=5)
+        self.assertNotAlmostEqual(float(result[2][2]), expected_22, places=5)
+        self.assertNotAlmostEqual(float(result[2][3]), expected_23, places=5)
+        self.assertNotAlmostEqual(float(result[3][0]), expected_30, places=5)
+        self.assertNotAlmostEqual(float(result[3][1]), expected_31, places=5)
+        self.assertNotAlmostEqual(float(result[3][2]), expected_32, places=5)
+        self.assertNotAlmostEqual(float(result[3][3]), expected_33, places=5)
+        self.assertAlmostEqual(float(result[0][0]), expected_00, places=1)
+        self.assertAlmostEqual(float(result[0][1]), expected_01, places=1)
+        self.assertAlmostEqual(float(result[0][2]), expected_02, places=1)
+        self.assertAlmostEqual(float(result[0][3]), expected_03, places=1)
+        self.assertAlmostEqual(float(result[1][0]), expected_10, places=1)
+        self.assertAlmostEqual(float(result[1][1]), expected_11, places=1)
+        self.assertAlmostEqual(float(result[1][2]), expected_12, places=1)
+        self.assertAlmostEqual(float(result[1][3]), expected_13, places=1)
+        self.assertAlmostEqual(float(result[2][0]), expected_20, places=1)
+        self.assertAlmostEqual(float(result[2][1]), expected_21, places=1)
+        self.assertAlmostEqual(float(result[2][2]), expected_22, places=1)
+        self.assertAlmostEqual(float(result[2][3]), expected_23, places=1)
+        self.assertAlmostEqual(float(result[3][0]), expected_30, places=1)
+        self.assertAlmostEqual(float(result[3][1]), expected_31, places=1)
+        self.assertAlmostEqual(float(result[3][2]), expected_32, places=1)
+        self.assertAlmostEqual(float(result[3][3]), expected_33, places=1)
 
     def test_vec2_arg_precision(self):
         values = (1.23, 2.34)
         expected = 2.64357712200722438922
 
         result = wp.length(wp.vec2d(*values))
-        self.assertAlmostEqual(result, expected, places=12)
+        self.assertIsInstance(result, wp.float64)
+        self.assertAlmostEqual(float(result), expected, places=12)
 
         result = wp.length(wp.vec2f(*values))
+        self.assertIsInstance(result, float)
         self.assertNotAlmostEqual(result, expected, places=12)
         self.assertAlmostEqual(result, expected, places=5)
 
         result = wp.length(wp.vec2h(*values))
-        self.assertNotAlmostEqual(result, expected, places=5)
-        self.assertAlmostEqual(result, expected, places=1)
+        self.assertIsInstance(result, wp.float16)
+        self.assertNotAlmostEqual(float(result), expected, places=5)
+        self.assertAlmostEqual(float(result), expected, places=1)
 
     def test_vec2_arg_overflow(self):
         values = (-1234567890, -1234567890)
@@ -481,15 +517,18 @@ class TestBuiltinsResolution(unittest.TestCase):
         expected = 4.34637780226247727455
 
         result = wp.length(wp.vec3d(*values))
-        self.assertAlmostEqual(result, expected, places=12)
+        self.assertIsInstance(result, wp.float64)
+        self.assertAlmostEqual(float(result), expected, places=12)
 
         result = wp.length(wp.vec3f(*values))
+        self.assertIsInstance(result, float)
         self.assertNotAlmostEqual(result, expected, places=12)
         self.assertAlmostEqual(result, expected, places=5)
 
         result = wp.length(wp.vec3h(*values))
-        self.assertNotAlmostEqual(result, expected, places=5)
-        self.assertAlmostEqual(result, expected, places=1)
+        self.assertIsInstance(result, wp.float16)
+        self.assertNotAlmostEqual(float(result), expected, places=5)
+        self.assertAlmostEqual(float(result), expected, places=1)
 
     def test_vec3_arg_overflow(self):
         values = (-1234567890, -1234567890, -1234567890)
@@ -509,15 +548,18 @@ class TestBuiltinsResolution(unittest.TestCase):
         expected = 6.29957141399317777086
 
         result = wp.length(wp.vec4d(*values))
-        self.assertAlmostEqual(result, expected, places=12)
+        self.assertIsInstance(result, wp.float64)
+        self.assertAlmostEqual(float(result), expected, places=12)
 
         result = wp.length(wp.vec4f(*values))
+        self.assertIsInstance(result, float)
         self.assertNotAlmostEqual(result, expected, places=12)
         self.assertAlmostEqual(result, expected, places=5)
 
         result = wp.length(wp.vec4h(*values))
-        self.assertNotAlmostEqual(result, expected, places=5)
-        self.assertAlmostEqual(result, expected, places=1)
+        self.assertIsInstance(result, wp.float16)
+        self.assertNotAlmostEqual(float(result), expected, places=5)
+        self.assertAlmostEqual(float(result), expected, places=1)
 
     def test_vec4_arg_overflow(self):
         values = (-1234567890, -1234567890, -1234567890, -1234567890)
@@ -538,15 +580,18 @@ class TestBuiltinsResolution(unittest.TestCase):
         expected = 14.91389999999999815827
 
         result = wp.dot(wp.vec2d(*a_values), wp.vec2d(*b_values))
-        self.assertAlmostEqual(result, expected, places=12)
+        self.assertIsInstance(result, wp.float64)
+        self.assertAlmostEqual(float(result), expected, places=12)
 
         result = wp.dot(wp.vec2f(*a_values), wp.vec2f(*b_values))
+        self.assertIsInstance(result, float)
         self.assertNotAlmostEqual(result, expected, places=12)
         self.assertAlmostEqual(result, expected, places=5)
 
         result = wp.dot(wp.vec2h(*a_values), wp.vec2h(*b_values))
-        self.assertNotAlmostEqual(result, expected, places=5)
-        self.assertAlmostEqual(result, expected, places=1)
+        self.assertIsInstance(result, wp.float16)
+        self.assertNotAlmostEqual(float(result), expected, places=5)
+        self.assertAlmostEqual(float(result), expected, places=1)
 
     def test_vec2_vec2_args_overflow(self):
         values = (-1234567890, -1234567890)
@@ -567,15 +612,18 @@ class TestBuiltinsResolution(unittest.TestCase):
         expected = 42.26760000000000161435
 
         result = wp.dot(wp.vec3d(*a_values), wp.vec3d(*b_values))
-        self.assertAlmostEqual(result, expected, places=12)
+        self.assertIsInstance(result, wp.float64)
+        self.assertAlmostEqual(float(result), expected, places=12)
 
         result = wp.dot(wp.vec3f(*a_values), wp.vec3f(*b_values))
+        self.assertIsInstance(result, float)
         self.assertNotAlmostEqual(result, expected, places=12)
         self.assertAlmostEqual(result, expected, places=5)
 
         result = wp.dot(wp.vec3h(*a_values), wp.vec3h(*b_values))
-        self.assertNotAlmostEqual(result, expected, places=5)
-        self.assertAlmostEqual(result, expected, places=1)
+        self.assertIsInstance(result, wp.float16)
+        self.assertNotAlmostEqual(float(result), expected, places=5)
+        self.assertAlmostEqual(float(result), expected, places=1)
 
     def test_vec3_vec3_args_overflow(self):
         values = (-1234567890, -1234567890, -1234567890)
@@ -596,15 +644,18 @@ class TestBuiltinsResolution(unittest.TestCase):
         expected = 90.64379999999999881766
 
         result = wp.dot(wp.vec4d(*a_values), wp.vec4d(*b_values))
-        self.assertAlmostEqual(result, expected, places=12)
+        self.assertIsInstance(result, wp.float64)
+        self.assertAlmostEqual(float(result), expected, places=12)
 
         result = wp.dot(wp.vec4f(*a_values), wp.vec4f(*b_values))
+        self.assertIsInstance(result, float)
         self.assertNotAlmostEqual(result, expected, places=12)
         self.assertAlmostEqual(result, expected, places=5)
 
         result = wp.dot(wp.vec4h(*a_values), wp.vec4h(*b_values))
-        self.assertNotAlmostEqual(result, expected, places=5)
-        self.assertAlmostEqual(result, expected, places=1)
+        self.assertIsInstance(result, wp.float16)
+        self.assertNotAlmostEqual(float(result), expected, places=5)
+        self.assertAlmostEqual(float(result), expected, places=1)
 
     def test_vec4_vec4_args_overflow(self):
         values = (-1234567890, -1234567890, -1234567890, -1234567890)
@@ -626,8 +677,8 @@ class TestBuiltinsResolution(unittest.TestCase):
         expected_y = 8.07300000000000039790
 
         result = wp.mul(wp.vec2d(*a_values), wp.float64(b_value))
-        self.assertAlmostEqual(result[0], expected_x, places=12)
-        self.assertAlmostEqual(result[1], expected_y, places=12)
+        self.assertAlmostEqual(float(result[0]), expected_x, places=12)
+        self.assertAlmostEqual(float(result[1]), expected_y, places=12)
 
         result = wp.mul(wp.vec2f(*a_values), wp.float32(b_value))
         self.assertNotAlmostEqual(result[0], expected_x, places=12)
@@ -636,10 +687,10 @@ class TestBuiltinsResolution(unittest.TestCase):
         self.assertAlmostEqual(result[1], expected_y, places=5)
 
         result = wp.mul(wp.vec2h(*a_values), wp.float16(b_value))
-        self.assertNotAlmostEqual(result[0], expected_x, places=5)
-        self.assertNotAlmostEqual(result[1], expected_y, places=5)
-        self.assertAlmostEqual(result[0], expected_x, places=1)
-        self.assertAlmostEqual(result[1], expected_y, places=1)
+        self.assertNotAlmostEqual(float(result[0]), expected_x, places=5)
+        self.assertNotAlmostEqual(float(result[1]), expected_y, places=5)
+        self.assertAlmostEqual(float(result[0]), expected_x, places=1)
+        self.assertAlmostEqual(float(result[1]), expected_y, places=1)
 
     def test_vec3_float_args_precision(self):
         a_values = (1.23, 2.34, 3.45)
@@ -649,9 +700,9 @@ class TestBuiltinsResolution(unittest.TestCase):
         expected_z = 15.73199999999999931788
 
         result = wp.mul(wp.vec3d(*a_values), wp.float64(b_value))
-        self.assertAlmostEqual(result[0], expected_x, places=12)
-        self.assertAlmostEqual(result[1], expected_y, places=12)
-        self.assertAlmostEqual(result[2], expected_z, places=12)
+        self.assertAlmostEqual(float(result[0]), expected_x, places=12)
+        self.assertAlmostEqual(float(result[1]), expected_y, places=12)
+        self.assertAlmostEqual(float(result[2]), expected_z, places=12)
 
         result = wp.mul(wp.vec3f(*a_values), wp.float32(b_value))
         self.assertNotAlmostEqual(result[0], expected_x, places=12)
@@ -662,12 +713,12 @@ class TestBuiltinsResolution(unittest.TestCase):
         self.assertAlmostEqual(result[2], expected_z, places=5)
 
         result = wp.mul(wp.vec3h(*a_values), wp.float16(b_value))
-        self.assertNotAlmostEqual(result[0], expected_x, places=5)
-        self.assertNotAlmostEqual(result[1], expected_y, places=5)
-        self.assertNotAlmostEqual(result[2], expected_z, places=5)
-        self.assertAlmostEqual(result[0], expected_x, places=1)
-        self.assertAlmostEqual(result[1], expected_y, places=1)
-        self.assertAlmostEqual(result[2], expected_z, places=1)
+        self.assertNotAlmostEqual(float(result[0]), expected_x, places=5)
+        self.assertNotAlmostEqual(float(result[1]), expected_y, places=5)
+        self.assertNotAlmostEqual(float(result[2]), expected_z, places=5)
+        self.assertAlmostEqual(float(result[0]), expected_x, places=1)
+        self.assertAlmostEqual(float(result[1]), expected_y, places=1)
+        self.assertAlmostEqual(float(result[2]), expected_z, places=1)
 
     def test_vec4_float_args_precision(self):
         a_values = (1.23, 2.34, 3.45, 4.56)
@@ -678,10 +729,10 @@ class TestBuiltinsResolution(unittest.TestCase):
         expected_w = 25.85519999999999640750
 
         result = wp.mul(wp.vec4d(*a_values), wp.float64(b_value))
-        self.assertAlmostEqual(result[0], expected_x, places=12)
-        self.assertAlmostEqual(result[1], expected_y, places=12)
-        self.assertAlmostEqual(result[2], expected_z, places=12)
-        self.assertAlmostEqual(result[3], expected_w, places=12)
+        self.assertAlmostEqual(float(result[0]), expected_x, places=12)
+        self.assertAlmostEqual(float(result[1]), expected_y, places=12)
+        self.assertAlmostEqual(float(result[2]), expected_z, places=12)
+        self.assertAlmostEqual(float(result[3]), expected_w, places=12)
 
         result = wp.mul(wp.vec4f(*a_values), wp.float32(b_value))
         self.assertNotAlmostEqual(result[0], expected_x, places=12)
@@ -694,14 +745,14 @@ class TestBuiltinsResolution(unittest.TestCase):
         self.assertAlmostEqual(result[3], expected_w, places=5)
 
         result = wp.mul(wp.vec4h(*a_values), wp.float16(b_value))
-        self.assertNotAlmostEqual(result[0], expected_x, places=5)
-        self.assertNotAlmostEqual(result[1], expected_y, places=5)
-        self.assertNotAlmostEqual(result[2], expected_z, places=5)
-        self.assertNotAlmostEqual(result[3], expected_w, places=5)
-        self.assertAlmostEqual(result[0], expected_x, places=1)
-        self.assertAlmostEqual(result[1], expected_y, places=1)
-        self.assertAlmostEqual(result[2], expected_z, places=1)
-        self.assertAlmostEqual(result[3], expected_w, places=1)
+        self.assertNotAlmostEqual(float(result[0]), expected_x, places=5)
+        self.assertNotAlmostEqual(float(result[1]), expected_y, places=5)
+        self.assertNotAlmostEqual(float(result[2]), expected_z, places=5)
+        self.assertNotAlmostEqual(float(result[3]), expected_w, places=5)
+        self.assertAlmostEqual(float(result[0]), expected_x, places=1)
+        self.assertAlmostEqual(float(result[1]), expected_y, places=1)
+        self.assertAlmostEqual(float(result[2]), expected_z, places=1)
+        self.assertAlmostEqual(float(result[3]), expected_w, places=1)
 
 
 for dtype in wp._src.types.int_types: